interactive_cnnvis/backend/requirements.txt
Flask
Flask-Cors
matplotlib
numpy 
opencv-python
pandas
Pillow
requests
scikit-learn
scipy
torch
torch-lucent
torchvision
tqdm